requirements.txt
nltk~=3.6.5
joblib~=1.1.0
pytrec-eval-terrier
numpy
pandas